requirements.txt
streamlit
requests
pandas
Pillow
python-dotenv
